+++ src/components/FileRepo/FileRepo.js
@@ -2,41 +2,38 @@ import * as React from 'react';
 import { compose } from 'recompose';
 import { injectState } from 'freactal';
 import { withTheme } from 'emotion-theming';
-import { css } from 'emotion';
-import styled from 'react-emotion';
-import { isObject, get } from 'lodash';
+import { isObject } from 'lodash';
 import { Trans } from 'react-i18next';
-import Spinner from 'react-spinkit';
 import FilterIcon from 'react-icons/lib/fa/filter';
 
 import { Arranger, CurrentSQON, Table, DetectNewVersion } from '@arranger/components/dist/Arranger';
 import { replaceSQON } from '@arranger/components/dist/SQONView/utils';
-import Query from '@arranger/components/dist/Query';
-
-// TODO: bringing beagle in through arrangerStyle seems to break the prod build...
-import '@arranger/components/public/themeStyles/beagle/beagle.css';
-// import arrangerStyle from './arrangerStyle';
 
 import SQONURL from 'components/SQONURL';
 import SaveQuery from 'components/ShareSaveQuery/SaveQuery';
 import ShareQuery from 'components/ShareSaveQuery/ShareQuery';
-import DownloadFileButton from 'components/FileRepo/DownloadFileButton';
 import FileRepoSidebar from 'components/FileRepoSidebar';
 import { FileRepoStats, FileRepoStatsQuery } from 'components/Stats';
 import ArrangerConnectionGuard from 'components/ArrangerConnectionGuard';
 import AggregationSidebar from 'components/FileRepo/AggregationSidebar';
-import { arrangerGqlRecompose } from 'services/arranger';
 import { Gen3UserProvider, getStudyIds } from 'services/gen3';
-
 import DownloadIcon from 'icons/DownloadIcon';
-
 import translateSQON from 'common/translateSQONValue';
 import { arrangerProjectId } from 'common/injectGlobals';
 import { withApi } from 'services/api';
 import { trackUserInteraction, TRACKING_EVENTS } from 'services/analyticsTracking';
-import Column from 'uikit/Column';
 import { FilterInput } from 'uikit/Input';
 import Row from 'uikit/Row';
+import {
+  ArrangerContainer,
+  TableContainer,
+  TableWrapper,
+  QuerySharingContainer,
+  ControlledIcon,
+  OpenIcon,
+  TableSpinner,
+} from './ui';
+import customTableColumns from './customTableColumns';
 
 const trackFileRepoInteraction = ({ label, ...eventData }) =>
   trackUserInteraction({
@@ -46,179 +43,10 @@ const trackFileRepoInteraction = ({ label, ...eventData }) =>
     ...(label && { label: isObject(label) ? JSON.stringify(label) : label }),
   });
 
-const ComposedSpinner = props => (
-  <Spinner
-    fadeIn="none"
-    name="circle"
-    color="#a9adc0"
-    style={{ width: 50, height: 50 }}
-    {...props}
-  />
-);
-
-const ControlledIcon = () => (
-  <img
-    src={require('../../assets/icon-controlled-access.svg')}
-    alt=""
-    css={`
-      width: 11px;
-      margin: auto;
-      display: block;
-    `}
-  />
-);
-
-const OpenIcon = () => (
-  <img
-    src={require('../../assets/icon-open-access.svg')}
-    alt=""
-    css={`
-      width: 10px;
-      margin: auto;
-      display: block;
-    `}
-  />
-);
-
 const customTableTypes = {
   access: ({ value }) =>
     typeof value !== 'boolean' ? `` : value ? <ControlledIcon /> : <OpenIcon />,
 };
-
-const customTableColumns = ({ theme, userProjectIds, loadingGen3User }) => [
-  {
-    index: 13,
-    content: {
-      accessor: 'kf_id',
-      Header: () => <DownloadIcon width={13} fill={theme.greyScale3} />,
-      Cell: withApi(({ value, api }) => (
-        <Query
-          renderError
-          api={arrangerGqlRecompose(api, 'TableRowStudyId')}
-          projectId={'june_13'}
-          shouldFetch={true}
-          query={`query ($sqon: JSON) {
-            file {
-              aggregations(filters: $sqon) {
-                participants__study__external_id {
-                  buckets {
-                    key
-                  }
-                }
-              }
-            }
-          }`}
-          variables={{
-            sqon: {
-              op: 'and',
-              content: [
-                {
-                  op: 'in',
-                  content: {
-                    field: 'kf_id',
-                    value: [value],
-                  },
-                },
-              ],
-            },
-          }}
-          render={({ loading: loadingQuery, data }) => {
-            const studyIdBucket = (get(
-              data,
-              'file.aggregations.participants__study__external_id.buckets',
-            ) || [])[0];
-            return (
-              <div
-                className={css`
-                  display: flex;
-                  align-items: center;
-                  justify-content: center;
-                `}
-              >
-                {studyIdBucket ? (
-                  userProjectIds.includes(studyIdBucket.key) ? (
-                    <DownloadFileButton kfId={value} />
-                  ) : (
-                    <ControlledIcon />
-                  )
-                ) : loadingQuery || loadingGen3User ? (
-                  <ComposedSpinner style={{ width: 20, height: 20 }} />
-                ) : (
-                  <ControlledIcon />
-                )}
-              </div>
-            );
-          }}
-        />
-      )),
-      width: 40,
-      sortable: false,
-      resizable: false,
-    },
-  },
-];
-
-const ArrangerContainer = styled(Row)`
-  .ReactTable .rt-thead .rt-th.-sort-desc,
-  .ReactTable .rt-thead .rt-td.-sort-desc {
-    box-shadow: inset 0 -3px 0 0 rgba(64, 76, 154, 0.7);
-  }
-
-  .ReactTable .rt-thead .rt-th.-sort-asc,
-  .ReactTable .rt-thead .rt-td.-sort-asc {
-    box-shadow: inset 0 3px 0 0 rgba(64, 76, 154, 0.7);
-  }
-
-  .tableToolbar {
-    border-left: solid 1px #e0e1e6;
-    border-right: solid 1px #e0e1e6;
-  }
-
-  & .tableToolbar {
-    color: ${({ theme }) => theme.greyScale9};
-    & .group .dropDownButtonContent {
-      color: ${({ theme }) => theme.greyScale9};
-    }
-    & .group button {
-      color: ${({ theme }) => theme.greyScale9};
-    }
-  }
-
-  .pagination-bottom .-pagination {
-    color: ${({ theme }) => theme.greyScale9};
-    select {
-      color: ${({ theme }) => theme.greyScale9};
-    }
-  }
-
-  div.sqon-view {
-    flex-grow: 1;
-  }
-`;
-
-const TableContainer = styled(Column)`
-  flex-grow: 1;
-  width: 580px;
-  padding: 30px;
-  position: relative;
-  height: 100%;
-  overflow-y: auto;
-`;
-
-const TableWrapper = styled(Column)`
-  min-height: 300px;
-  flex: 1;
-  & .ReactTable {
-    min-height: 1px;
-  }
-`;
-
-const QuerySharingContainer = styled(Row)`
-  border-style: solid;
-  border-color: ${({ theme }) => theme.borderGrey};
-  border-width: 1px 1px 1px 0;
-  background: ${({ theme }) => theme.backgroundGrey};
-`;
 
 const FileRepo = compose(injectState, withTheme, withApi)(
   ({
@@ -243,7 +71,7 @@ const FileRepo = compose(injectState, withTheme, withApi)(
                 {connectionError ? (
                   `Unable to connect to the file repo, please try again later`
                 ) : (
-                  <ComposedSpinner />
+                  <TableSpinner />
                 )}
               </div>
             ) : (
